Massive refactoring of parsing
We now use an enum to represent the elements and make rules clearer.
The parser should be able to accept all valid nvmbers now, and also
able to reject any invalid ones.

This change also moves to using i64 as the main data type to represent
nvmbers internally, making it possible to represent negative values.